requirements.txt
openai>=1.0.0
anthropic>=0.5.0
google-generativeai>=0.2.0
mistralai>=0.0.1
groq>=0.0.1
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
tqdm>=4.65.0
wandb>=0.15.0
torch>=2.0.0
transformers>=4.30.0
tenacity>=8.2.0
requests>=2.31.0
scipy>=1.7.0
gym>=0.21.0
